% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-reading.R
\docType{package}
\name{accelerate-package}
\alias{accelerate-package}
\alias{accelerate}
\title{accelerate-package}
\description{
\tabular{ll}{
Package: \tab accelerate\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-12-08\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\details{
Read in and summarize accelerometer data
}
\examples{
NULL


}
\references{
\url{http://en.wikipedia.org/wiki/accelerometry}
}
\seealso{
\code{\link{read.actigraph}}
}
\author{
Kenn Konstabel \email{kenn.konstabel AT tai.ee}
}
\keyword{package}
