% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-summarize.R
\name{aciPlot}
\alias{aciPlot}
\title{Plot time activity categories by 30-min segments}
\usage{
aciPlot(
  D,
  cutname = "Evenson",
  cats = c("Sedentary", "Light", "Moderate", "Vigorous"),
  plot = TRUE,
  select = Wkdy \%in\% 0:6,
  ...
)
}
\arguments{
\item{D}{summarized data (using 30 mins as period with summary.acc)}

\item{cutname}{Name of cut points}

\item{cats}{Cutoff categories as a vector (eg., sedentary, light, moderate)}

\item{plot}{if TRUE, the results will be plotted using plotAci}

\item{select}{select}

\item{...}{...}
}
\value{
cdplot.acc
}
\description{
aciPlot
}
