% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-internal.R
\name{actigraphmode}
\alias{actigraphmode}
\title{actigraphmode}
\usage{
actigraphmode(MODE, MODEL)
}
\arguments{
\item{MODE}{mode}

\item{MODEL}{model}
}
\value{
actigraphmode
}
\description{
actigraphmode
}
\keyword{internal}
