% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-process.R
\name{as.zoo.acc}
\alias{as.zoo.acc}
\title{Convert an "acc" object to zoo}
\usage{
\method{as.zoo}{acc}(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{extra dots passed to as.zoo}
}
\value{
zoo obj
}
\description{
as.zoo.acc
}
