% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-process.R
\name{BOUTS}
\alias{BOUTS}
\title{Bouts}
\usage{
BOUTS(Counts, Epoch, BoutLength = 5, CUT, Threshold = 0.8)
}
\arguments{
\item{Counts}{counts}

\item{Epoch}{epoch}

\item{BoutLength}{bout length}

\item{CUT}{cut}

\item{Threshold}{.8}
}
\value{
bouts
}
\description{
Bouts
}
