% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-process.R
\name{CNTSTATS.3dnx}
\alias{CNTSTATS.3dnx}
\title{3dnx cntstats}
\usage{
CNTSTATS.3dnx(
  x,
  Epoch,
  cutoffs = Cutoffs,
  return.MVPA = TRUE,
  return.bouts = TRUE,
  return.rel = TRUE,
  return.raw = TRUE,
  rel.factor = 1,
  reintegrate.to = NULL
)
}
\arguments{
\item{x}{x}

\item{Epoch}{epoch}

\item{cutoffs}{cutoffs}

\item{return.MVPA}{TRUE}

\item{return.bouts}{TRUE}

\item{return.rel}{TRUE}

\item{return.raw}{TRUE}

\item{rel.factor}{1}

\item{reintegrate.to}{NULL}
}
\value{
3dnx cntstats
}
\description{
3dnx cntstats
}
