% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-process.R
\name{CNTSTATS2}
\alias{CNTSTATS2}
\title{cntstats2}
\usage{
CNTSTATS2(x, Epoch, cutoffs = Cutoffs.Evenson, ...)
}
\arguments{
\item{x}{x}

\item{Epoch}{epoch}

\item{cutoffs}{evenson}

\item{...}{dots}
}
\value{
cntstats
}
\description{
cntstats2
}
