% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-internal.R
\name{Curry}
\alias{Curry}
\alias{[.function}
\title{Curry with brackets}
\usage{
Curry(FUN, ...)

\method{[}{`function`}(FUN, ...)
}
\arguments{
\item{FUN}{function}

\item{...}{dots}
}
\value{
function
}
\description{
Curry with brackets
}
