% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-cutoffs.R
\docType{data}
\name{Cutoffs.3dnx}
\alias{Cutoffs.3dnx}
\title{3dnx cutoffs}
\format{
An object of class \code{list} of length 1.
}
\usage{
Cutoffs.3dnx
}
\description{
3dnx cutoffs
}
\keyword{datasets}
