% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-cutoffs.R
\docType{data}
\name{Cutoffs.Evenson}
\alias{Cutoffs.Evenson}
\title{Evenson cutoffs}
\format{
An object of class \code{list} of length 1.
}
\usage{
Cutoffs.Evenson
}
\description{
Evenson cutoffs
}
\keyword{datasets}
