% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-cutoffs.R
\docType{data}
\name{Cutoffs.Freedson98}
\alias{Cutoffs.Freedson98}
\title{Freedson 1998 cutoffs}
\format{
An object of class \code{list} of length 1.
}
\usage{
Cutoffs.Freedson98
}
\description{
Freedson 1998 cutoffs
}
\keyword{datasets}
