% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-cutoffs.R
\docType{data}
\name{Cutoffs.FreedsonVM}
\alias{Cutoffs.FreedsonVM}
\title{Freedson VM cutoffs}
\format{
An object of class \code{list} of length 1.
}
\usage{
Cutoffs.FreedsonVM
}
\description{
Freedson VM cutoffs
}
\keyword{datasets}
