% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-cutoffs.R
\docType{data}
\name{Cutoffs.Helena}
\alias{Cutoffs.Helena}
\title{Helena cutoffs}
\format{
An object of class \code{list} of length 1.
}
\usage{
Cutoffs.Helena
}
\description{
Helena cutoffs
}
\keyword{datasets}
