% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-cutoffs.R
\docType{data}
\name{Cutoffs.Pate}
\alias{Cutoffs.Pate}
\title{Pate cutoffs}
\format{
An object of class \code{list} of length 1.
}
\usage{
Cutoffs.Pate
}
\description{
Pate cutoffs
}
\keyword{datasets}
