% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-cutoffs.R
\docType{data}
\name{Cutoffs.Pate2}
\alias{Cutoffs.Pate2}
\title{Pate2 cutoffs}
\format{
An object of class \code{list} of length 1.
}
\usage{
Cutoffs.Pate2
}
\description{
Pate2 cutoffs
}
\keyword{datasets}
