% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-cutoffs.R
\docType{data}
\name{Cutoffs.Puyau}
\alias{Cutoffs.Puyau}
\title{Puyau cutoffs}
\format{
An object of class \code{list} of length 1.
}
\usage{
Cutoffs.Puyau
}
\description{
Puyau cutoffs
}
\keyword{datasets}
