% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-cutoffs.R
\docType{data}
\name{Cutoffs.Reilly}
\alias{Cutoffs.Reilly}
\title{Reilly cutoffs}
\format{
An object of class \code{list} of length 1.
}
\usage{
Cutoffs.Reilly
}
\description{
Reilly cutoffs
}
\keyword{datasets}
