% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-cutoffs.R
\docType{data}
\name{Cutoffs.Sirard}
\alias{Cutoffs.Sirard}
\title{Sirard cutoffs}
\format{
An object of class \code{list} of length 1.
}
\usage{
Cutoffs.Sirard
}
\description{
Sirard cutoffs
}
\keyword{datasets}
