% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-cutoffs.R
\docType{data}
\name{Cutoffs.Troiano}
\alias{Cutoffs.Troiano}
\title{Troiano cutoffs}
\format{
An object of class \code{list} of length 1.
}
\usage{
Cutoffs.Troiano
}
\description{
Troiano cutoffs
}
\keyword{datasets}
