% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-process.R
\name{cuttable}
\alias{cuttable}
\title{Cuttable}
\usage{
cuttable(x, breaks, include.lowest = TRUE, right = FALSE, ...)
}
\arguments{
\item{x}{x}

\item{breaks}{breaks}

\item{right}{FALSE}

\item{...}{dots}

\item{include.lowerst}{TRUE}
}
\value{
bincount
}
\description{
Cuttable
}
\keyword{internal}
