% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-process.R
\name{delete.zeros}
\alias{delete.zeros}
\title{Delete consecutive zeros}
\usage{
delete.zeros(x, period = 20)
}
\arguments{
\item{x}{the file to load}

\item{period}{minimum length}
}
\value{
x
}
\description{
Consecutive zeros (of at least 'period' minutes) 
are replaced by NA
}
