% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-feedback.R
\name{do.feedback}
\alias{do.feedback}
\title{do.feedback}
\usage{
do.feedback(
  FILE,
  DIR,
  TSDIR,
  PDF = TRUE,
  browse = FALSE,
  PARAMS = "Estonian",
  chg.locale = FALSE,
  FUN = NULL
)
}
\arguments{
\item{FILE}{filename}

\item{DIR}{dir}

\item{TSDIR}{a folder where to put the resulting pdf files}

\item{PDF}{TRUE}

\item{browse}{FALSE}

\item{PARAMS}{params}

\item{chg.locale}{chg.locale}

\item{FUN}{NULL}
}
\description{
do feedback
}
