% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-summarize.R
\name{file.ext}
\alias{file.ext}
\title{Get file extension}
\usage{
file.ext(FN)
}
\arguments{
\item{FN}{file name}
}
\value{
file extension (the part after the last "." in the file name)
}
\description{
file.ext
}
