% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-process.R
\name{FILTER}
\alias{FILTER}
\title{Filter out implausible values}
\usage{
FILTER(X, HD, ERR, maxsteps = 500, maxcnts = 15000, minhr = 60, maxhr = 220)
}
\arguments{
\item{X}{data part of an acc object}

\item{HD}{header data}

\item{ERR}{env to send err...}

\item{maxsteps}{maxsteps}

\item{maxcnts}{maxcnts default 15000}

\item{minhr}{minhr default 60}

\item{maxhr}{maxhr default 220}
}
\value{
filtered X
}
\description{
Filter out implausible values
}
