% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-internal.R
\name{get.counts}
\alias{get.counts}
\title{Get counts}
\usage{
get.counts(x, method = "first")
}
\arguments{
\item{x}{'acc object'}

\item{method}{method to get counts}
}
\value{
counts
}
\description{
Get counts
}
\keyword{internal}
