% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-reading.R
\name{getNAME}
\alias{getNAME}
\title{Get character ID from file name}
\usage{
getNAME(x, noblanks = TRUE)
}
\arguments{
\item{x}{file name (character)}

\item{noblanks}{TRUE (blanks will be deleted)}
}
\value{
ID (as character)
}
\description{
Get ID from file name. File extension will be cut and 
everything else written with leading capital letter.
For example, 'Elijah Bailey.dat' becomes 'ElijahBailey'
}
