% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-reading.R
\name{getRex}
\alias{getRex}
\title{Get ID from file name (2)}
\usage{
getRex(x)
}
\arguments{
\item{x}{file name (character)}
}
\value{
ID (as character)
}
\description{
Get ID from file name: cut off the extension 
and treat the rest as ID. 
For example, 'Elijah Bailey.dat' becomes 'Elijah Bailey'
}
