% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-feedback.R
\name{gplotsTextplotMatrix}
\alias{gplotsTextplotMatrix}
\title{gplotsTextplotMatrix}
\usage{
gplotsTextplotMatrix(
  object,
  halign = c("center", "left", "right"),
  valign = c("center", "top", "bottom"),
  cex,
  cmar = 2,
  rmar = 0.5,
  show.rownames = TRUE,
  show.colnames = TRUE,
  hadj = 1,
  vadj = 1,
  mar = c(1, 1, 4, 1) + 0.1,
  col.data = graphics::par("col"),
  col.rownames = graphics::par("col"),
  col.colnames = graphics::par("col"),
  ...
)
}
\arguments{
\item{object}{object}

\item{halign}{halign}

\item{valign}{valign}

\item{cex}{cex}

\item{cmar}{2}

\item{rmar}{0.5}

\item{show.rownames}{TRUE}

\item{show.colnames}{TRUE}

\item{hadj}{1}

\item{vadj}{1}

\item{mar}{1 1 4 1}

\item{col.data}{par('col')}

\item{col.rownames}{par('col')}

\item{col.colnames}{par('col')}

\item{...}{...}
}
\description{
gplotsTextplotMatrix
}
