% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-reading.R
\name{guess.dates}
\alias{guess.dates}
\title{Guess dates}
\usage{
guess.dates(
  d1,
  t1,
  d2,
  t2,
  START = ISOdate(2007, 8, 31),
  END = ISOdate(2008, 7, 1)
)
}
\arguments{
\item{d1}{date 1}

\item{t1}{time 1}

\item{d2}{date 2}

\item{t2}{time 2}

\item{START}{start date of study}

\item{END}{end date of study}
}
\value{
an object (list) of class "acc"
}
\description{
Date format guessing used in read.actigraph.dat and read.csa
}
\details{
For reading in older actigraph files in "dat" format.
The main shortcoming of this format is ambiguity of date format -
it is dependent on computer used to download the data. Therefore
the newer format (agd) should be used when possible.
}
\seealso{
\code{\link{read.actigraph.csv}}
}
