% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-process.R
\name{HRSTATS2}
\alias{HRSTATS2}
\title{hrstats2}
\usage{
HRSTATS2(x, Epoch, cutoffs = Cutoffs.Evenson, ...)
}
\arguments{
\item{x}{x}

\item{Epoch}{epoch}

\item{cutoffs}{evenson}

\item{...}{dots}
}
\value{
hrstats
}
\description{
hrstats2
}
