% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-internal.R
\name{iCurry}
\alias{iCurry}
\title{iCurry}
\usage{
iCurry(FUN, x, ...)
}
\arguments{
\item{FUN}{function}

\item{x}{x}

\item{...}{dots}
}
\value{
function with na.rm=T if possible
}
\description{
iCurry
}
\keyword{internal}
