% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-summarize.R
\name{load.acc.dir}
\alias{load.acc.dir}
\title{load.acc.dir}
\usage{
load.acc.dir(folder, ...)
}
\arguments{
\item{folder}{folder}

\item{...}{...}
}
\value{
load.acc.dir
}
\description{
load.acc.dir
}
