% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-process.R
\name{load.acc}
\alias{load.acc}
\title{load.acc}
\usage{
load.acc(file, DIR, get.id = identity, ...)
}
\arguments{
\item{file}{file}

\item{DIR}{dir}

\item{get.id}{get.id}

\item{...}{dots}
}
\value{
acc obj
}
\description{
load.acc
}
