% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-cutoffs.R
\name{make.cutoffs}
\alias{make.cutoffs}
\title{Make cutoffs}
\usage{
make.cutoffs(Name, Epoch, ...)
}
\arguments{
\item{Name}{name}

\item{Epoch}{epoch}

\item{...}{dots}
}
\value{
cutoffs
}
\description{
Make cutoffs
}
