% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-process.R
\name{nofilter}
\alias{nofilter}
\title{Nofilter}
\usage{
nofilter(X, ...)
}
\arguments{
\item{X}{acc object}

\item{...}{discarded}
}
\value{
X
}
\description{
Filler function when no filter is used
}
