% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-wm.R
\name{pa_choi}
\alias{pa_choi}
\title{Choi algorithm (detection of nonwear time)}
\usage{
pa_choi(ct, frame = 90, streamFrame = 30, allowanceFrame = 2, talkie = FALSE)
}
\arguments{
\item{ct}{A vector of counts}

\item{frame}{90}

\item{streamFrame}{30}

\item{allowanceFrame}{2}

\item{talkie}{FALSE}
}
\value{
X
}
\description{
Choi algorithm (detection of nonwear time)
}
\details{
This is the "workhorse" function for wmChoi and is borrowed from PhysicalActivity:::marking (not exported from PhysicalActivity, thus included here as a copy rather than as a dependency)
}
\author{
Leena Choi <leena.choi@Vanderbilt.Edu>, Zhouwen Liu <zhouwen.liu@Vanderbilt.Edu>, Charles E. Matthews <Charles.Matthews2@nih.gov>, and Maciej S. Buchowski <maciej.buchowski@Vanderbilt.Edu>
}
