% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-plots.R
\name{pdfplot}
\alias{pdfplot}
\title{pdfplot}
\usage{
pdfplot(x, filename, MFROW = c(3, 2), PAPER = "a4", ...)
}
\arguments{
\item{x}{x}

\item{filename}{filename}

\item{MFROW}{mfrow}

\item{PAPER}{a4}

\item{...}{dots}
}
\value{
nothing useful
}
\description{
pdfplot
}
