% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-plots.R
\name{plot.acc}
\alias{plot.acc}
\title{Plot an "acc" object}
\usage{
\method{plot}{acc}(
  x,
  BY = NULL,
  cutoffs = NULL,
  c.factor = 3,
  threshold = 60 * 6,
  fileintitle = TRUE,
  ...
)
}
\arguments{
\item{x}{x}

\item{BY}{BY}

\item{cutoffs}{cutoffs}

\item{c.factor}{3}

\item{threshold}{360}

\item{fileintitle}{TRUE}

\item{...}{dots}
}
\value{
nothing useful
}
\description{
plot.acc
}
