% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-process.R
\name{print.acc}
\alias{print.acc}
\title{print.acc}
\usage{
\method{print}{acc}(x, FORMAT = "\%A, \%d. \%B \%Y \%H:\%M:\%S", ...)
}
\arguments{
\item{x}{acc object}

\item{FORMAT}{date format}

\item{...}{dots}
}
\value{
nothing useful
}
\description{
print.acc
}
