% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-reading.R
\name{process.acc}
\alias{process.acc}
\title{process.acc}
\usage{
process.acc(x, .filter = FILTER, preprocess = delete.zeros, get.id = NULL)
}
\arguments{
\item{x}{x}

\item{.filter}{.filter}

\item{preprocess}{preprocess}

\item{get.id}{get.id}
}
\value{
process.acc
}
\description{
process.acc
}
