% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-summarize.R
\name{process.folder}
\alias{process.folder}
\title{Apply a set of procedures to all files in a folder}
\usage{
process.folder(
  FOLDER,
  FUN,
  FUN2 = NULL,
  SAVE1 = NULL,
  SAVE2 = NULL,
  FINALLY = NULL,
  LIST = NULL,
  EXT = NULL,
  PATTERN = "",
  ALL = FALSE,
  FULL = TRUE,
  VERBOSE = TRUE,
  FILES = NULL,
  TIME = TRUE,
  ...
)
}
\arguments{
\item{FOLDER}{(character) name of a folder}

\item{FUN}{a function}

\item{FUN2}{a funciton}

\item{SAVE1}{a function}

\item{SAVE2}{a function}

\item{FINALLY}{a function}

\item{LIST}{a vector (numeric or logical) of files to include}

\item{EXT}{file extension(s) to include}

\item{PATTERN}{(character) pattern of file names - passed to list.files}

\item{ALL}{(boolean) include all files in the folder?}

\item{FULL}{(boolean) use full names}

\item{VERBOSE}{(boolean) report all the superfluous details?}

\item{FILES}{(character vector) a set of files possibly from different folders; if not NULL then the FOLDER argument is ignored}

\item{TIME}{(boolean) sets timing on/off}

\item{...}{...}
}
\value{
process.folder
}
\description{
process.folder
}
\details{
This applies a set of rules to all (or some: as per LIST) files in a folder (with the same extensions, specified in EXT). FUN is first applied to each file (using try), and then FUN2. Errors are collected and reported as the component $errors in the result. Finally, the function given in FINALLY is applied to combine the results (using do.call). For saving the results, specify a function in SAVE. Finally, the results are timed if TIME==TRUE. See read.actigraph.folder for an example.
}
