% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-reading.R
\name{read.3dnx}
\alias{read.3dnx}
\title{Read in a 3DNX (TM) file}
\usage{
read.3dnx(
  file,
  DIR,
  METHOD = rowSums,
  default.epoch = 5,
  ask.if.no.epoch = FALSE,
  get.id = getID,
  preprocess = delete.zeros,
  method = "sum",
  ...
)
}
\arguments{
\item{file}{the file to read}

\item{DIR}{folder - added to 'file' but ignored if 'file' contains '/'}

\item{METHOD}{rowSums}

\item{default.epoch}{what to do if epoch is missing from file}

\item{ask.if.no.epoch}{FALSE}

\item{get.id}{getID}

\item{preprocess}{delete.zeros}

\item{method}{"sum"}

\item{...}{args 2b passed}
}
\value{
an object (list) of class "acc"
}
\description{
Reads in a 3DNX file and
returns an object of class "acc" that can be further
processed with summary.acc etc.
}
\details{
For reading in 3dnx files
}
\seealso{
\code{\link{read.actigraph.csv}}
}
