% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-reading.R
\name{read.actigraph.csv}
\alias{read.actigraph.csv}
\title{Read in actigraph *.csv file}
\usage{
read.actigraph.csv(
  file,
  DIR = "data",
  get.id = getID,
  preprocess = delete.zeros,
  .filter = FILTER,
  ...
)
}
\arguments{
\item{file}{the file to read}

\item{DIR}{folder - added to 'file' but can be missing}

\item{get.id}{- function to extract ID from file name, eg. getID}

\item{preprocess}{- function to preprocess data - either delete.zeros or reintegrate - or a custom function}

\item{.filter}{- function to filter out implausible values}

\item{...}{extra parameters are silently ignored}
}
\value{
an object (list) of class "acc"
}
\description{
Reads in an Actigraph (GT1M, ActiTrainer) csv file and
returns an object of class "acc" that can be further
processed with summary.acc etc.
}
\details{
For reading in older actigraph files in "dat" format.
The main shortcoming of this format is ambiguity of date format -
it is dependent on computer used to download the data. Therefore
the newer format (agd) should be used when possible.
}
\seealso{
\code{\link{read.actigraph.csv}}
}
