% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-summarize.R
\name{read.actigraph.folder}
\alias{read.actigraph.folder}
\title{read.actigraph.folder}
\usage{
read.actigraph.folder(
  FOLDER,
  LIST = NULL,
  RIT = 15,
  CUTOFFS = Cutoffs.T1,
  BY. = "DSTday",
  STATS. = c(CNTSTATS, STEPSTATS, HRSTATS),
  EXT. = c("dat", "csv", "agd"),
  ...
)
}
\arguments{
\item{FOLDER}{name of the folder}

\item{LIST}{which files from the folder to read in (could be numeric or logical index)}

\item{RIT}{reintegrate to common epoch (default is 15)}

\item{CUTOFFS}{which cutoffs to use (default to a long list used in Idefics study: Cutoffs.T1)}

\item{BY.}{period to use as a unit of analysis: defults to "DSTday" (calendar day but using daylight saving time changes when relevant i.e period is usually 24 hours but can be 23 or 25); could be anything like "30 mins" or "2 days"}

\item{STATS.}{defaults to c(CNTSTATS, STEPSTATS, HRSTATS) - that is - calculate statistics for activity counts, steps, and heart rate}

\item{EXT.}{which file extensions to use; defaults to c("dat", "csv", "agd")}

\item{...}{...}
}
\value{
read.actigraph.folder
}
\description{
read.actigraph.folder
}
