% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-reading.R
\name{read.actigraph.header}
\alias{read.actigraph.header}
\title{Process an actigraph file header}
\usage{
read.actigraph.header(
  filename,
  dateformat = c("month", "day", "year"),
  id.prefix = "",
  format = TRUE,
  h = FALSE,
  Way = "ActiGraph",
  return.l = FALSE
)
}
\arguments{
\item{filename}{the file to read}

\item{dateformat}{- the order of "month" "day" and "year" in date string given as a character vector e.g c("month", "date", "year")}

\item{id.prefix}{- prefix to be added to the id}

\item{format}{TRUE}

\item{h}{FALSE}

\item{Way}{"ActiGraph" or "CSA"}

\item{return.l}{FALSE}
}
\value{
a list of processed header data
}
\description{
Reads in the header part of an actigraph 'dat' file and 
does the necessary text processing. Used in read.actigraph.dat.
}
\details{
For processing the headers of older actigraph files in "dat" format.
}
\seealso{
\code{\link{read.actigraph.dat}}
}
