% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-reading.R
\name{read.actigraph}
\alias{read.actigraph}
\title{read.actigraph}
\usage{
read.actigraph(.., ...)
}
\arguments{
\item{..}{a file}

\item{...}{args to be passed}
}
\value{
an object of class 'acc'
}
\description{
read.actigraph
}
