% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-reading.R
\name{read.agd}
\alias{read.agd}
\title{read.agd}
\usage{
read.agd(
  file,
  DIR,
  get.id = getID,
  getinside = FALSE,
  limit = NULL,
  query = NULL,
  ...
)
}
\arguments{
\item{file}{file}

\item{DIR}{folder name (but you can include this in file name too)}

\item{get.id}{get.id}

\item{getinside}{getinside}

\item{limit}{limit to the number of lines to be read from the file}

\item{query}{either NULL or a valid SQL query}

\item{...}{...}
}
\value{
read.agd
}
\description{
read.agd
}
