% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-reading.R
\name{read.agd_old}
\alias{read.agd_old}
\title{read.agd_old}
\usage{
read.agd_old(file, DIR, get.id = getID, justso = FALSE, ...)
}
\arguments{
\item{file}{file}

\item{DIR}{folder name (but you can include this in file name too)}

\item{get.id}{get.id}

\item{justso}{justso}

\item{...}{...}
}
\value{
read.agd
}
\description{
read.agd_old
}
