% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-reading.R
\name{read.csa}
\alias{read.csa}
\title{Read in a CSA (older Actigraph models ) file}
\usage{
read.csa(file, DIR, get.id = identity, ...)
}
\arguments{
\item{file}{the file to read}

\item{DIR}{folder - added to 'file' but ignored if 'file' contains '/'}

\item{get.id}{- function to extract ID from file name, eg. getID}

\item{...}{any extra parameters are silently ignored (check if this can be deleted)}
}
\value{
an object (list) of class "acc"
}
\description{
Reads in a CSA file and returns an object of class "acc" that can be 
further processed with summary.acc etc.
}
\details{
For reading in older CSA actigraph files. The 
format is very similar to the 'dat' format used with GT1M and ActiTrainer
and has the same shortcomings.
}
\seealso{
\code{\link{read.actigraph.csv}}
}
