% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-process.R
\name{reintegrate}
\alias{reintegrate}
\title{Reintegrate}
\usage{
reintegrate(
  x,
  TO = NULL,
  FACTOR = NULL,
  after = delete.zeros,
  fractional = warning,
  ...
)
}
\arguments{
\item{x}{acc object}

\item{TO}{target epoch}

\item{FACTOR}{factor}

\item{after}{function to be applied afterwards}

\item{fractional}{function to be called when fractional epoch occurs: normally either warning or error, possibly print or cat}

\item{...}{passed to 'after'}
}
\value{
x
}
\description{
Reintegrate to another (longer) epoch
}
