% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-summarize.R
\name{summarize}
\alias{summarize}
\title{summarize}
\usage{
summarize(
  A,
  condition = val.time >= 360,
  no.days = 3,
  keep = "all",
  head = c("File", "ID", "Epoch", "Period"),
  subset.only = FALSE
)
}
\arguments{
\item{A}{A data frame with accelerometer results (such as the result of summary.acc or read.actigraph.folder)}

\item{condition}{inclusion condition (column names in table A can be used, e.g., "val.time >= 480")}

\item{no.days}{number of days}

\item{keep}{not implemented yet}

\item{head}{names of variables to include in the result}

\item{subset.only}{subset.only}
}
\value{
summarize
}
\description{
summarize
}
