% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-summarize2.R
\name{summarize2}
\alias{summarize2}
\title{Summarize2 - by weekdays and weekends}
\usage{
summarize2(
  x,
  reqdays = c(wd = 2, we = 1),
  cutoffs,
  excl.rel = TRUE,
  epoch,
  minhrs,
  include2 = c("Period", "Length", "Wkdy", "avg.cpm", "tot.cnts", "val.time", "permax")
)
}
\arguments{
\item{x}{a data frame, result from read.actigraph.folder or similar}

\item{reqdays}{vector or list with elements named wd and we - required days}

\item{cutoffs}{character (name of the set of cutoffs to be used)}

\item{excl.rel}{exclude "relative" times in activity categories}

\item{epoch}{epoch}

\item{minhrs}{minimum required hours}

\item{include2}{include2}
}
\value{
a df
}
\description{
Cf summarize
}
\details{
Like summarize but see ortega et al in PLOS!
}
\seealso{
\code{\link{summarize}}
}
