% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-internal.R
\docType{data}
\name{validEpochs}
\alias{validEpochs}
\alias{Wkdy}
\alias{val.time}
\title{Internal accelerate functions}
\format{
An object of class \code{character} of length 1.
}
\usage{
validEpochs
}
\description{
Infernal variables
}
\keyword{datasets}
