% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerate-wm.R
\name{wmChoi}
\alias{wmChoi}
\title{Choi wearing marking}
\usage{
wmChoi(ACC, COL = "Activity1", set.na = TRUE, add.wmcol = FALSE, ...)
}
\arguments{
\item{ACC}{object of class "acc"}

\item{COL}{column name in ACC$X that contains counts (defaults to "Activity1")}

\item{set.na}{whether to set nonwear time as NA}

\item{add.wmcol}{whether to add an extra column to ACC$X for wearing marking}

\item{...}{extra params passed to choi}
}
\value{
X
}
\description{
Choi wearing marking
}
\details{
The algorithm implemented in this function is described in 100. Choi L, Liu Z, Matthews CE, Buchowski MS. Validation of accelerometer wear and nonwear time classification algorithm. Med Sci Sports Exerc. 2011 Feb;43(2):357-64. doi: 10.1249/MSS.0b013e3181ed61a3. This function does the preparatory job; the actual algoritm is in `pa_choi`.
}
